Implement Authorization using Authorities,Roles

+++ section9/springsecsection9/src/main/java/com/eazybytes/repository/CardsRepository.java
@@ -2,14 +2,15 @@ package com.eazybytes.repository;
 
 import java.util.List;
 
+import com.eazybytes.model.Loans;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 import com.eazybytes.model.Cards;
 
 @Repository
-public interface CardsRepository extends CrudRepository<Cards, Integer> {
+public interface CardsRepository extends CrudRepository<Cards, Long> {
 	
-	List<Cards> findByCustomerId(int customerId);
+	List<Cards> findByCustomerId(long customerId);
 
 }
