+++ section9/springsecsection9/src/main/java/com/eazybytes/repository/LoanRepository.java
@@ -8,8 +8,8 @@ import org.springframework.stereotype.Repository;
 import com.eazybytes.model.Loans;
 
 @Repository
-public interface LoanRepository extends CrudRepository<Loans, Integer> {
+public interface LoanRepository extends CrudRepository<Loans, Long> {
 	
-	List<Loans> findByCustomerIdOrderByStartDtDesc(int customerId);
+	List<Loans> findByCustomerIdOrderByStartDtDesc(long customerId);
 
 }
